Moved OfficialTemplatePicker to common.
