+++ web-dashboard/src/components/common/officialTemplatePicker/OfficialTemplatePicker.tsx
