+++ web-dashboard/src/components/common/officialTemplatePicker/index.ts
